Partial refactor of databases/queries
