+++ manager-rs/dauth-service/src/local/database/mod.rs
@@ -1,0 +1,8 @@
+pub mod user_infos;
+pub mod kseafs;
+pub mod general;
+pub mod flood_vectors;
+pub mod backup_users;
+pub mod backup_networks;
+pub mod auth_vectors;
+pub mod key_shares;
